Rename to something else to prevent too many confusing changes
https://coveord.atlassian.net/browse/JSUI-2674
