+++ src/controllers/CategoryFacetQueryController.ts
@@ -1,75 +1,86 @@
 import { CategoryFacet } from '../ui/CategoryFacet/CategoryFacet';
-import { Assert } from '../misc/Assert';
 import { QueryBuilder } from '../ui/Base/QueryBuilder';
-import { IFacetRequest, IFacetRequestValue } from '../rest/Facet/FacetRequest';
-import { QueryEvents } from '../events/QueryEvents';
-import { CategoryFacetValue } from '../ui/CategoryFacet/CategoryFacetValues/CategoryFacetValue';
-import { FacetValueState } from '../rest/Facet/FacetValueState';
+import { ICategoryFacetRequest } from '../rest/CategoryFacetRequest';
+import { IGroupByRequest } from '../rest/GroupByRequest';
+import { AllowedValuesPatternType } from '../rest/AllowedValuesPatternType';
+import { IGroupByValue } from '../rest/GroupByValue';
+import { Utils } from '../utils/Utils';
 
+// TODO: remove this file and any reference to it
 export class CategoryFacetQueryController {
-  private numberOfValuesToRequest: number;
-  private freezeFacetOrder = false;
+  constructor(private categoryFacet: CategoryFacet) {}
 
-  constructor(private facet: CategoryFacet) {
-    this.resetNumberOfValuesToRequest();
-    this.resetFreezeCurrentValuesDuringQuery();
+  public putCategoryFacetInQueryBuilder(queryBuilder: QueryBuilder, path: string[], maximumNumberOfValues: number): number {
+    const positionInQuery = queryBuilder.categoryFacets.length;
+    this.addQueryFilter(queryBuilder, path);
+    this.addCategoryFacetRequest(queryBuilder, path, maximumNumberOfValues);
+    return positionInQuery;
   }
 
-  private resetFreezeCurrentValuesDuringQuery() {
-    this.facet.bind.onRootElement(QueryEvents.duringQuery, () => {
-      this.freezeFacetOrder = false;
+  public async searchFacetValues(value: string, numberOfValues: number): Promise<IGroupByValue[]> {
+    const lastQuery = { ...this.categoryFacet.queryController.getLastQuery() };
+
+    const groupByRequest: IGroupByRequest = {
+      allowedValues: [this.getAllowedValuesPattern(value)],
+      allowedValuesPatternType: AllowedValuesPatternType.Wildcards,
+      maximumNumberOfValues: numberOfValues,
+      field: this.categoryFacet.options.field as string,
+      sortCriteria: 'occurrences',
+      injectionDepth: this.categoryFacet.options.injectionDepth
+    };
+
+    lastQuery.groupBy = [groupByRequest];
+    lastQuery.categoryFacets.splice(this.categoryFacet.positionInQuery, 1);
+    const results = await this.categoryFacet.queryController.getEndpoint().search(lastQuery);
+
+    const sortByNumberOfResultsThenPathLength = (firstGroupByValue: IGroupByValue, secondGroupByValue: IGroupByValue) => {
+      if (firstGroupByValue.numberOfResults == secondGroupByValue.numberOfResults) {
+        return firstGroupByValue.value.length - secondGroupByValue.value.length;
+      }
+      return secondGroupByValue.numberOfResults - firstGroupByValue.numberOfResults;
+    };
+
+    return results.groupByResults[0].values.sort(sortByNumberOfResultsThenPathLength);
+  }
+
+  public addDebugGroupBy(queryBuilder: QueryBuilder, value: string) {
+    queryBuilder.groupByRequests.push({
+      field: this.categoryFacet.options.field as string,
+      allowedValues: [`.*${Utils.escapeRegexCharacter(value)}.*`],
+      allowedValuesPatternType: AllowedValuesPatternType.Regex
     });
   }
 
-  public increaseNumberOfValuesToRequest(additionalNumberOfValues: number) {
-    this.numberOfValuesToRequest += additionalNumberOfValues;
+  private shouldAddFilterToQuery(path: string[]) {
+    return path.length != 0 && !Utils.arrayEqual(path, this.categoryFacet.options.basePath);
   }
 
-  public resetNumberOfValuesToRequest() {
-    this.numberOfValuesToRequest = this.facet.options.numberOfValues;
-  }
-
-  public putFacetIntoQueryBuilder(queryBuilder: QueryBuilder) {
-    Assert.exists(queryBuilder);
-
-    queryBuilder.facetRequests.push(this.facetRequest);
-    if (this.freezeFacetOrder) {
-      queryBuilder.facetOptions.freezeFacetOrder = true;
+  private addQueryFilter(queryBuilder: QueryBuilder, path: string[]) {
+    if (this.shouldAddFilterToQuery(path)) {
+      queryBuilder.advancedExpression.addFieldExpression(this.categoryFacet.options.field as string, '==', [
+        path.join(this.categoryFacet.options.delimitingCharacter)
+      ]);
     }
   }
 
-  public get facetRequest(): IFacetRequest {
-    return {
-      facetId: this.facet.options.id,
-      field: this.facet.fieldName,
-      type: this.facet.facetType,
-      currentValues: this.currentValues,
-      numberOfValues: this.numberOfValues,
-      delimitingCharacter: this.facet.options.delimitingCharacter,
-      isFieldExpanded: this.numberOfValuesToRequest > this.facet.options.numberOfValues
+  private addCategoryFacetRequest(queryBuilder: QueryBuilder, path: string[], maximumNumberOfValues: number) {
+    const categoryFacetsRequest: ICategoryFacetRequest = {
+      field: this.categoryFacet.options.field as string,
+      path,
+      injectionDepth: this.categoryFacet.options.injectionDepth,
+      maximumNumberOfValues,
+      delimitingCharacter: this.categoryFacet.options.delimitingCharacter
     };
+    queryBuilder.categoryFacets.push(categoryFacetsRequest);
   }
 
-  public get currentValues(): IFacetRequestValue[] {
-    return this.facet.values.allFacetValues.map(requestValue => this.buildRequestValue(requestValue));
-  }
+  private getAllowedValuesPattern(value: string) {
+    const basePath = this.categoryFacet.options.basePath;
+    const delimiter = this.categoryFacet.options.delimitingCharacter;
 
-  private buildRequestValue(facetValue: CategoryFacetValue): IFacetRequestValue {
-    return {
-      value: facetValue.value,
-      state: facetValue.state,
-      preventAutoSelect: facetValue.preventAutoSelect,
-      children: facetValue.children.map(requestValue => this.buildRequestValue(requestValue)),
-      retrieveChildren: this.shouldRetrieveChildren(facetValue),
-      retrieveCount: 3, // TODO: Move numberOfValuesToRequest to every child values
+    if (Utils.isNonEmptyArray(basePath)) {
+      return `${basePath.join(delimiter)}${delimiter}*${value}*`;
     }
-  }
-
-  private shouldRetrieveChildren(facetValue: CategoryFacetValue) {
-    return !facetValue.children.length && facetValue.state === FacetValueState.selected;
-  }
-
-  private get numberOfValues() {
-    return this.numberOfValuesToRequest;
+    return `*${value}*`;
   }
 }
